requirements.txt
# Core ML/DL
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
Pillow>=10.0.0

# Medical Image Processing
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git
connected-components-3d>=3.10.0

# Federated Learning
flwr>=1.6.0
flwr-datasets>=0.0.2

# Web API
flask>=3.0.0
flask-cors>=4.0.0
werkzeug>=3.0.0

# XAI (Explainable AI)
grad-cam>=1.4.8
captum>=0.6.0
lime>=0.2.0.1
shap>=0.43.0

# Data handling
pandas>=2.0.0
tqdm>=4.66.0

# Differential Privacy (optional for FL)
opacus>=1.4.0

# Visualization
matplotlib>=3.7.0

# PyQt for GUI
PyQt5>=5.15.9
